+++ src/main.rs
@@ -1,7 +1,4 @@
-use nix::sys::signal::{self as nix_signal, Signal};
-use nix::sys::wait::WaitStatus;
-use nix::{self, unistd::Pid};
-use std::process::{exit, Command, Stdio};
+use std::process::{Command, Stdio};
 use std::sync::{Arc, Barrier, Mutex};
 use std::thread;
 
@@ -68,61 +65,10 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     // use handle_signal
     let procs_2 = Arc::clone(&procs);
-
-    let check_child_terminated = thread::Builder::new()
-        .name(String::from(format!("check child terminated")))
-        .spawn(move || {
-            loop {
-                // Waiting for the end of any one child process
-                match nix::sys::wait::waitpid(
-                    Pid::from_raw(-1),
-                    Some(nix::sys::wait::WaitPidFlag::WNOHANG),
-                ) {
-                    Ok(exit_status) => match exit_status {
-                        WaitStatus::Exited(pid, code) => {
-                            procs.lock().unwrap().retain(|p| {
-                                let child_id = p.lock().unwrap().child.id() as i32;
-                                Pid::from_raw(child_id) != pid
-                            });
-
-                            // If the child process dies, send SIGTERM to all child processes
-                            for proc in procs.lock().unwrap().iter() {
-                                let proc = proc.lock().unwrap();
-                                let child_id = proc.child.id();
-
-                                log::output(
-                                    "system",
-                                    &format!(
-                                        "sending SIGTERM for {} at pid {}",
-                                        &proc.name, &child_id
-                                    ),
-                                    padding
-                                );
-                                nix_signal::kill(Pid::from_raw(child_id as i32), Signal::SIGTERM)
-                                    .unwrap();
-                            }
-                            log::output("system", &format!("exit {}", &code), padding);
-                            // close loop (thread finished)
-                            exit(code);
-                        }
-                        _ => (),
-                    },
-                    Err(e) => {
-                        if let nix::Error::Sys(nix::errno::Errno::ECHILD) = e {
-                            // close loop (thread finished)
-                            exit(0);
-                        }
-                    }
-                };
-            }
-        })?;
-    proc_handles.push(check_child_terminated);
+    proc_handles.push(process::check_child_terminated(procs, padding));
 
     let procs = Arc::clone(&procs_2);
-    let handle_signal = thread::Builder::new()
-        .name(String::from("handling signal"))
-        .spawn(move || signal::handle_signal(procs, padding).expect("fail to handle signal"))?;
-    proc_handles.push(handle_signal);
+    proc_handles.push(signal::handle_signal(procs, padding));
 
     for handle in proc_handles {
         handle.join().expect("failed join");
